Move Last Active Rubric Name to local storage
