+++ teachertool/src/services/indexedDbService.ts
@@ -6,10 +6,6 @@ import { Rubric } from "../types/rubric";
 const teacherToolDbName = "makecode-project-insights";
 const dbVersion = 1;
 const rubricsStoreName = "rubrics";
-const metadataStoreName = "metadata";
-const metadataKeys = {
-    lastActiveRubricKey: "lastActiveRubricName",
-};
 
 type MetadataEntry = { key: string; value: any };
 
@@ -21,7 +17,6 @@ class TeacherToolDb {
         this.db = await openDB(teacherToolDbName, dbVersion, {
             upgrade(db) {
                 db.createObjectStore(rubricsStoreName, { keyPath: "name" });
-                db.createObjectStore(metadataStoreName, { keyPath: "key" });
             },
         });
     }
@@ -64,27 +59,6 @@ class TeacherToolDb {
         }
     }
 
-    private async getMetadataEntryAsync(key: string): Promise<MetadataEntry | undefined> {
-        return this.getAsync<MetadataEntry>(metadataStoreName, key);
-    }
-
-    private async setMetadataEntryAsync(key: string, value: any): Promise<void> {
-        return this.setAsync<MetadataEntry>(metadataStoreName, { key, value });
-    }
-
-    private async deleteMetadataEntryAsync(key: string): Promise<void> {
-        return this.deleteAsync(metadataStoreName, key);
-    }
-
-    public async getLastActiveRubricNameAsync(): Promise<string | undefined> {
-        const metadataEntry = await this.getMetadataEntryAsync(metadataKeys.lastActiveRubricKey);
-        return metadataEntry?.value;
-    }
-
-    public saveLastActiveRubricNameAsync(name: string): Promise<void> {
-        return this.setMetadataEntryAsync(metadataKeys.lastActiveRubricKey, name);
-    }
-
     public getRubric(name: string): Promise<Rubric | undefined> {
         return this.getAsync<Rubric>(rubricsStoreName, name);
     }
@@ -104,11 +78,10 @@ const getDb = (async () => {
     return db;
 })();
 
-export async function getLastActiveRubricAsync(): Promise<Rubric | undefined> {
+export async function getRubricFromIndexedDbAsync(name: string): Promise<Rubric | undefined> {
     const db = await getDb;
 
     let rubric: Rubric | undefined = undefined;
-    const name = await db.getLastActiveRubricNameAsync();
     if (name) {
         rubric = await db.getRubric(name);
     }
@@ -116,13 +89,12 @@ export async function getLastActiveRubricAsync(): Promise<Rubric | undefined> {
     return rubric;
 }
 
-export async function saveRubricAsync(rubric: Rubric) {
+export async function saveRubricToIndexedDbAsync(rubric: Rubric) {
     const db = await getDb;
     await db.saveRubric(rubric);
-    await db.saveLastActiveRubricNameAsync(rubric.name);
 }
 
-export async function deleteRubricAsync(name: string) {
+export async function deleteRubricFromIndexedDbAsync(name: string) {
     const db = await getDb;
     await db.deleteRubric(name);
 }
