+++ teachertool/src/services/localStorage.ts
@@ -3,6 +3,7 @@ import { logError } from "./loggingService";
 
 const KEY_PREFIX = "teachertool";
 const AUTORUN_KEY = [KEY_PREFIX, "autorun"].join("/");
+const LAST_ACTIVE_RUBRIC_KEY = [KEY_PREFIX, "lastActiveRubric"].join("/");
 
 function getValue(key: string, defaultValue?: string): string | undefined {
     return localStorage.getItem(key) || defaultValue;
@@ -16,7 +17,7 @@ function delValue(key: string) {
     localStorage.removeItem(key);
 }
 
-function getAutorun(): boolean {
+export function getAutorun(): boolean {
     try {
         return getValue(AUTORUN_KEY, "false") === "true";
     } catch (e) {
@@ -25,7 +26,7 @@ function getAutorun(): boolean {
     }
 }
 
-function setAutorun(autorun: boolean) {
+export function setAutorun(autorun: boolean) {
     try {
         setValue(AUTORUN_KEY, autorun.toString());
     } catch (e) {
@@ -33,4 +34,19 @@ function setAutorun(autorun: boolean) {
     }
 }
 
-export { getAutorun, setAutorun };
+export function getLastActiveRubricName(): string {
+    try {
+        return getValue(LAST_ACTIVE_RUBRIC_KEY) ?? "";
+    } catch (e) {
+        logError(ErrorCode.localStorageReadError, e);
+        return "";
+    }
+}
+
+export function setLastActiveRubricName(name: string) {
+    try {
+        setValue(LAST_ACTIVE_RUBRIC_KEY, name);
+    } catch (e) {
+        logError(ErrorCode.localStorageWriteError, e);
+    }
+}
